chore(cdn/mathjax): split worker util
